+++ CHANGELOG.md
@@ -5,6 +5,12 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ### [Unreleased]
 
+### Changed
+- Renamed JNDI properties 
+  - `stream.processing.max.event.retries`
+  - `stream.processing.retry.delay.milliseconds`
+  - `stream.processing.retry.delay.multiplier`
+
 ### [17.105.0-M16]  - 2026-03-24
 ### Added
 - Notification-based event linking and publishing via CDI events, enabled via JNDI: